title_parse now handles a wide variety of inputs
this commit closes issue #48
----------------------------

all unit tests in file are passing.  some random testing will be needed
to identify patterns of naming not yet anticipated or seen, however the
rules in this routine should be working a very large majority of the
time.

on failure, this routine should return ('',input_given), however all
cases of dirty input have not been tested to ensure it always fails
gracefully